Se implementaron los dto, los interface y se organizaron todas las clases con su respectivo generador de id

+++ src/main/java/com/BasesDeDatos/Proyecto/repositorio/envio/EnvioRepo.java
@@ -3,5 +3,5 @@ package com.BasesDeDatos.Proyecto.repositorio.envio;
 import com.BasesDeDatos.Proyecto.model.entidades.envios.Envio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnvioRepo extends JpaRepository<Envio, Long> {
+public interface EnvioRepo extends JpaRepository<Envio, Integer> {
 }
