+++ src/main/java/com/BasesDeDatos/Proyecto/repositorio/producto/DisenoEstampadoRepo.java
@@ -3,6 +3,6 @@ package com.BasesDeDatos.Proyecto.repositorio.producto;
 import com.BasesDeDatos.Proyecto.model.entidades.producto.DisenoEstampado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DisenoEstampadoRepo  extends JpaRepository<DisenoEstampado, Long> {
+public interface DisenoEstampadoRepo  extends JpaRepository<DisenoEstampado, Integer> {
 
 }
