+++ src/main/java/com/BasesDeDatos/Proyecto/repositorio/producto/MaterialDeConstruccionRepo.java
@@ -3,5 +3,5 @@ package com.BasesDeDatos.Proyecto.repositorio.producto;
 import com.BasesDeDatos.Proyecto.model.entidades.producto.MaterialDeConstruccion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MaterialDeConstruccionRepo  extends JpaRepository<MaterialDeConstruccion, Long> {
+public interface MaterialDeConstruccionRepo  extends JpaRepository<MaterialDeConstruccion, Integer> {
 }
