+++ src/main/java/com/BasesDeDatos/Proyecto/repositorio/producto/ModoDeDisparoRepo.java
@@ -3,5 +3,5 @@ package com.BasesDeDatos.Proyecto.repositorio.producto;
 import com.BasesDeDatos.Proyecto.model.entidades.producto.ModoDeDisparo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ModoDeDisparoRepo extends JpaRepository<ModoDeDisparo, Long> {
+public interface ModoDeDisparoRepo extends JpaRepository<ModoDeDisparo, Integer> {
 }
