+++ src/main/java/com/BasesDeDatos/Proyecto/repositorio/producto/MunicionRepo.java
@@ -3,5 +3,5 @@ package com.BasesDeDatos.Proyecto.repositorio.producto;
 import com.BasesDeDatos.Proyecto.model.entidades.producto.Municion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MunicionRepo  extends JpaRepository<Municion, Long> {
+public interface MunicionRepo  extends JpaRepository<Municion, Integer> {
 }
