+++ src/main/java/com/BasesDeDatos/Proyecto/repositorio/producto/PromocionRepo.java
@@ -3,5 +3,5 @@ package com.BasesDeDatos.Proyecto.repositorio.producto;
 import com.BasesDeDatos.Proyecto.model.entidades.producto.Promocion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PromocionRepo extends JpaRepository<Promocion, Long> {
+public interface PromocionRepo extends JpaRepository<Promocion, Integer> {
 }
