+++ src/main/java/com/BasesDeDatos/Proyecto/repositorio/producto/TipoDeArmaRepo.java
@@ -3,5 +3,5 @@ package com.BasesDeDatos.Proyecto.repositorio.producto;
 import com.BasesDeDatos.Proyecto.model.entidades.producto.TipoDeArma;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TipoDeArmaRepo extends JpaRepository<TipoDeArma, Long> {
+public interface TipoDeArmaRepo extends JpaRepository<TipoDeArma, Integer> {
 }
