+++ src/main/java/com/BasesDeDatos/Proyecto/repositorio/ubicacion/DepartamentoRepo.java
@@ -3,5 +3,5 @@ package com.BasesDeDatos.Proyecto.repositorio.ubicacion;
 import com.BasesDeDatos.Proyecto.model.entidades.ubicacion.Departamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartamentoRepo extends JpaRepository<Departamento, Long> {
+public interface DepartamentoRepo extends JpaRepository<Departamento, Integer> {
 }
