+++ src/main/java/com/BasesDeDatos/Proyecto/repositorio/ubicacion/DireccionRepo.java
@@ -3,5 +3,5 @@ package com.BasesDeDatos.Proyecto.repositorio.ubicacion;
 import com.BasesDeDatos.Proyecto.model.entidades.ubicacion.Direccion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DireccionRepo extends JpaRepository<Direccion, Long> {
+public interface DireccionRepo extends JpaRepository<Direccion, Integer> {
 }
