+++ src/main/java/com/BasesDeDatos/Proyecto/repositorio/vendedores/ComisionVentasRepo.java
@@ -3,5 +3,5 @@ package com.BasesDeDatos.Proyecto.repositorio.vendedores;
 import com.BasesDeDatos.Proyecto.model.entidades.vendedor.ComisionVentas;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ComisionVentasRepo extends JpaRepository<ComisionVentas, Long> {
+public interface ComisionVentasRepo extends JpaRepository<ComisionVentas, Integer> {
 }
