+++ src/main/java/com/BasesDeDatos/Proyecto/repositorio/vendedores/HistorialRankingRepo.java
@@ -3,5 +3,5 @@ package com.BasesDeDatos.Proyecto.repositorio.vendedores;
 import com.BasesDeDatos.Proyecto.model.entidades.vendedor.HistorialRanking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HistorialRankingRepo  extends JpaRepository<HistorialRanking, Long> {
+public interface HistorialRankingRepo  extends JpaRepository<HistorialRanking, Integer> {
 }
