+++ src/main/java/com/BasesDeDatos/Proyecto/repositorio/vendedores/PagoComisionesRepo.java
@@ -3,6 +3,6 @@ package com.BasesDeDatos.Proyecto.repositorio.vendedores;
 import com.BasesDeDatos.Proyecto.model.entidades.vendedor.PagoComisiones;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PagoComisionesRepo extends JpaRepository<PagoComisiones, Long>   {
+public interface PagoComisionesRepo extends JpaRepository<PagoComisiones, Integer>   {
 
 }
