+++ src/main/java/com/BasesDeDatos/Proyecto/repositorio/vendedores/RankingRepo.java
@@ -3,5 +3,5 @@ package com.BasesDeDatos.Proyecto.repositorio.vendedores;
 import com.BasesDeDatos.Proyecto.model.entidades.vendedor.Ranking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RankingRepo  extends JpaRepository<Ranking, Long> {
+public interface RankingRepo  extends JpaRepository<Ranking, Integer> {
 }
