+++ src/main/java/com/BasesDeDatos/Proyecto/repositorio/vendedores/VendedorRepo.java
@@ -3,5 +3,5 @@ package com.BasesDeDatos.Proyecto.repositorio.vendedores;
 import com.BasesDeDatos.Proyecto.model.entidades.vendedor.Vendedor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VendedorRepo  extends JpaRepository<Vendedor, Long> {
+public interface VendedorRepo  extends JpaRepository<Vendedor, Integer> {
 }
