+++ src/main/java/com/BasesDeDatos/Proyecto/repositorio/ventas/VentaRepo.java
@@ -3,5 +3,5 @@ package com.BasesDeDatos.Proyecto.repositorio.ventas;
 import com.BasesDeDatos.Proyecto.model.entidades.ventas.Venta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VentaRepo  extends JpaRepository<Venta, Long> {
+public interface VentaRepo  extends JpaRepository<Venta, Integer> {
 }
